+++ CHANGELOG.md
@@ -35,6 +35,7 @@ Please make sure to add your changes to the appropriate categories:
 - Added `Filter<f32>` quadrature map for `Nco<T>`
 - Added `Filter<T>` waveform maps for square and sawtooth oscillators
 - Added `Filter<T>` waveform maps for triangle and pulse oscillators
+- Added `Filter<T>` quadratic-phase map for `Chirp<T>`
 
 ### Changed